Correct link to extension in chrome web store (#127)
There were 3 links within the readme that all mistakenly had a slash at the end, such that they did NOT go to the extension page. Instead they went to the front page of the store.

Please note also that the same problem exists in the "about" text for the repo, shown on the right when viewing the front page of the repo. It seems that's not something I can propose a PR for, so I hope you will change that. That was actually the first link I clicked, which failed. Then I looked for and found these others.